Refactor stocktake error handling
